requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
altgraph==0.17.3
anyio==3.6.2
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asn1crypto==0.24.0
async-generator==1.10
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.1.0
autopep8==2.0.2
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
bitarray==2.7.3
bleach==6.0.0
blinker==1.6.1
blis==0.7.9
Brlapi==0.6.6
cachetools==5.3.0
catalogue==2.0.8
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
colorama==0.3.7
command-not-found==0.3
confection==0.0.4
cryptography==2.1.4
cupshelpers==1.0
cycler==0.11.0
cymem==2.0.7
dacite==1.8.1
dask==2022.2.0
debugpy==1.6.3
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distro-info===0.18ubuntu0.18.04.1
docx2txt==0.8
easyocr==1.6.2
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl#sha256=0964370218b7e1672a30ac50d72cdc6b16f7c867496f1d60925691188f4d2510
entrypoints==0.4
exceptiongroup==1.0.0rc9
fake-useragent==1.1.3
fastjsonschema==2.16.2
ffmpeg-python==0.2.0
filelock==3.9.0
flet==0.6.2
flet-core==0.6.2
fonttools==4.37.4
frozenlist==1.3.3
fsspec==2023.1.0
future==0.18.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
h11==0.14.0
htmlmin==0.1.12
httpcore==0.16.3
httplib2==0.9.2
httpx==0.23.3
huggingface-hub==0.12.1
idna==3.4
ImageHash==4.3.1
imageio==2.28.1
importlib-metadata==4.12.0
importlib-resources==5.9.0
ipykernel==6.16.2
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
iso3166==2.1.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.16.0
jupyter-server==1.24.0
jupyter_client==7.3.5
jupyter_core==4.12.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kazam==1.4.5
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.4.4
langchain==0.0.27
langcodes==3.3.0
language-selector==0.1
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
lazypredict==0.2.12
lightgbm==3.3.5
line-profiler==4.0.3
locket==1.0.0
louis==3.5.0
lux==0.5.1
lux-api==0.5.1
lux-widget==0.1.11
lxml==4.9.2
macaroonbakery==1.1.3
mahotas==1.4.13
Mako==1.0.7
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matlab-kernel==0.17.1
matplotlib==3.5.3
matplotlib-inline==0.1.6
mdurl==0.1.2
memory-profiler==0.60.0
metakernel==0.29.4
mistune==2.0.5
more-itertools==9.1.0
multidict==6.0.4
multimethod==1.9.1
murmurhash==1.0.9
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.6.0
nest-asyncio==1.5.5
netifaces==0.10.4
networkx==2.6.3
ninja==1.11.1
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauth==1.0.1
oauthlib==3.2.2
olefile==0.45.1
openai==0.27.2
openai-whisper==20230124
opencv-python-headless==4.5.4.60
openshot-qt==2.4.1
outcome==1.2.0
packaging==23.1
pandas==1.3.5
pandas-profiling==3.6.6
pandocfilters==1.5.0
parso==0.8.3
partd==1.4.0
pathy==0.10.1
patsy==0.5.3
pdfreader==0.1.12
pexpect==4.8.0
phik==0.12.3
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
plotly==5.15.0
polars==0.16.7
ppscore==1.3.0
preshed==3.0.8
prometheus-client==0.17.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.2
ptyprocess==0.7.0
pyarrow==11.0.0
pycairo==1.16.2
pyclipper==1.3.0.post4
pycodestyle==2.10.0
pycparser==2.21
pycrypto==2.6.1
pycryptodome==3.17
pycups==1.9.73
pydantic==1.10.7
pydeck==0.8.0
Pygments==2.13.0
PyGObject==3.26.1
pyinstaller==5.11.0
pyinstaller-hooks-contrib==2023.3
pymacaroons==0.13.0
Pympler==1.0.1
PyNaCl==1.1.2
pyparsing==3.0.9
PyPDF2==2.12.1
pyperclip==1.8.2
pyRFC3339==1.0
pyrsistent==0.18.1
PySimpleGUI==4.60.4
PySocks==1.7.1
pytesseract==0.3.10
python-apt==1.6.5+ubuntu0.7
python-bidi==0.4.2
python-dateutil==2.8.2
python-debian==0.1.32
python-dotenv==0.21.1
pytz==2018.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.25
PyYAML==6.0
pyzmq==24.0.1
QtPy==2.3.0
quickda==0.2.2
regex==2022.10.31
repath==0.9.0
reportlab==3.4.0
requests==2.28.2
requests-unixsocket==0.1.5
rfc3986==1.5.0
rich==13.3.3
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
scour==0.36
screen-resolution-extra==0.0.0
seaborn==0.12.2
SecretStorage==2.3.1
selenium==4.5.0
Send2Trash==1.8.2
sentencepiece==0.1.99
sh==1.14.3
shapely==2.0.1
simplejson==3.13.2
six==1.11.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.5.2
spacy-legacy==3.0.12
spacy-loggers==1.0.4
SQLAlchemy==2.0.7
srsly==2.4.6
statsmodels==0.13.5
streamlit==1.21.0
system-service==0.3
systemd-python==234
tangled-up-in-unicode==0.2.0
tenacity==8.2.2
terminado==0.17.1
thinc==8.1.10
threadpoolctl==3.1.0
tifffile==2021.11.2
tinycss2==1.2.1
tokenizers==0.13.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.65.0
traitlets==5.4.0
transformers==4.26.1
trio==0.22.0
trio-websocket==0.9.2
typeguard==2.13.3
typer==0.7.0
typing_extensions==4.4.0
tzdata==2023.3
tzlocal==4.3
ubuntu-advantage-tools==8001
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.26.12
usb-creator==0.3.3
validators==0.20.0
virtualenv==15.1.0
visions==0.7.5
wadllib==1.3.2
wasabi==1.1.1
watchdog==2.3.1
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.5.1
websockets==10.4
widgetsnbextension==4.0.7
wordcloud==1.9.2
wsproto==1.2.0
wurlitzer==3.0.3
xgboost==1.6.2
xkit==0.0.0
yarl==1.8.2
ydata-profiling==4.2.0
yellowbrick==1.5
zipp==3.8.1
zope.interface==4.3.2
